refactor: extract cleaner
