+++ idgen/delegated_test.go
@@ -92,76 +92,3 @@ func TestDelegatedRemove(t *testing.T) {
 		t.Fatal("unexpected count")
 	}
 }
-
-func TestDelegatedClean(t *testing.T) {
-	listener, err := net.Listen("tcp", ":0")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	bidg := NewBalancedIdentityGenerator()
-
-	mux := http.NewServeMux()
-	mux.HandleFunc("/idgen/add", func(w http.ResponseWriter, r *http.Request) {
-		pk, _ := bidg.AddBalanced()
-		b, _ := crypto.MarshalPrivateKey(pk)
-		json.NewEncoder(w).Encode(base64.StdEncoding.EncodeToString(b))
-	})
-	mux.HandleFunc("/idgen/remove", func(w http.ResponseWriter, r *http.Request) {
-		dec := json.NewDecoder(r.Body)
-		var b64 string
-		dec.Decode(&b64)
-		bytes, _ := base64.StdEncoding.DecodeString(b64)
-		pk, _ := crypto.UnmarshalPrivateKey(bytes)
-		err = bidg.Remove(pk)
-		if err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-		w.WriteHeader(http.StatusNoContent)
-	})
-
-	go http.Serve(listener, mux)
-	defer listener.Close()
-
-	count := bidg.Count()
-	if count != 0 {
-		t.Fatal("unexpected count")
-	}
-
-	didg := NewDelegatedIDGeneratorCleaner("http://" + listener.Addr().String())
-	_, err = didg.AddBalanced()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	pk, err := didg.AddBalanced()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	count = bidg.Count()
-	if count != 2 {
-		t.Fatal("unexpected count")
-	}
-
-	err = didg.Remove(pk)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	count = bidg.Count()
-	if count != 1 {
-		t.Fatal("unexpected count")
-	}
-
-	err = didg.Clean()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	count = bidg.Count()
-	if count != 0 {
-		t.Fatal("unexpected count")
-	}
-}
